+++ src/libtrx/game/ui/elements/scrollable_area.c
@@ -34,9 +34,9 @@ static void M_DownArrow(const UI_SCROLLABLE *const s)
     UI_Spacer(0.0f, TR_VERSION == 2 ? 6.0f : 4.0f);
 }
 
-void UI_BeginScrollableArea(const UI_SCROLLABLE *const s)
+void UI_BeginScrollableArea(const UI_SCROLLABLE *const s, const bool force_show)
 {
-    if (s->vis_items < s->max_items) {
+    if (s->vis_items < s->max_items || force_show) {
         UI_BeginStackEx((UI_STACK_SETTINGS) {
             .orientation = UI_STACK_VERTICAL,
             .align = { .h = UI_STACK_H_ALIGN_SPAN },
@@ -45,9 +45,9 @@ void UI_BeginScrollableArea(const UI_SCROLLABLE *const s)
     }
 }
 
-void UI_EndScrollableArea(const UI_SCROLLABLE *const s)
+void UI_EndScrollableArea(const UI_SCROLLABLE *const s, const bool force_show)
 {
-    if (s->vis_items < s->max_items) {
+    if (s->vis_items < s->max_items || force_show) {
         M_DownArrow(s);
         UI_EndStack();
     }