+++ src/libtrx/include/libtrx/game/ui/elements/scrollable_area.h
@@ -3,5 +3,5 @@
 #include "../common.h"
 #include "../scrollable.h"
 
-void UI_BeginScrollableArea(const UI_SCROLLABLE *s);
-void UI_EndScrollableArea(const UI_SCROLLABLE *s);
+void UI_BeginScrollableArea(const UI_SCROLLABLE *s, bool force_show);
+void UI_EndScrollableArea(const UI_SCROLLABLE *s, bool force_show);