+++ src/tr1/game/inventory_ring/vars.h
@@ -34,5 +34,6 @@ extern INVENTORY_ITEM g_InvRing_Item_Details;
 extern INVENTORY_ITEM g_InvRing_Item_Sound;
 extern INVENTORY_ITEM g_InvRing_Item_Controls;
 extern INVENTORY_ITEM g_InvRing_Item_LarasHome;
+extern INVENTORY_ITEM g_InvRing_Item_NatlasPDA;
 
 extern CAMERA_INFO g_InvRing_OldCamera;