+++ src/tr1/game/option/option_gameplay.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <libtrx/game/inventory_ring/types.h>
+
+void Option_Gameplay_Control(INVENTORY_ITEM *inv_item, bool is_busy);
+void Option_Gameplay_Draw(INVENTORY_ITEM *inv_item);
+void Option_Gameplay_Shutdown(void);