+++ src/tr2/game/ui/dialogs/graphic_settings.c
@@ -3,41 +3,41 @@
 #include <libtrx/config.h>
 
 static const UI_SETTINGS_ENUM_ENTRY m_CameraModeEnumEntries[] = {
-    { CAMERA_MODE_TR1, GS_ID(DETAIL_CAMERA_MODE_TR1) },
-    { CAMERA_MODE_TR2, GS_ID(DETAIL_CAMERA_MODE_TR2) },
+    { CAMERA_MODE_TR1, GS_ID(GRAPHIC_SETTINGS_CAMERA_MODE_TR1) },
+    { CAMERA_MODE_TR2, GS_ID(GRAPHIC_SETTINGS_CAMERA_MODE_TR2) },
     { -1, nullptr },
 };
 
 static const UI_SETTINGS_ENUM_ENTRY m_TextureFilterEnumEntries[] = {
     // clang-format off
     { GFX_TF_NN, GS_ID(MISC_OFF), },
-    { GFX_TF_BILINEAR, GS_ID(DETAIL_BILINEAR), },
+    { GFX_TF_BILINEAR, GS_ID(GRAPHIC_SETTINGS_BILINEAR), },
     { -1, nullptr, },
     // clang-format on
 };
 
 static const UI_SETTINGS_ENUM_ENTRY m_LightingContrastEnumEntries[] = {
     // clang-format off
-    { LIGHTING_CONTRAST_LOW, GS_ID(DETAIL_LIGHTING_CONTRAST_LOW), },
-    { LIGHTING_CONTRAST_MEDIUM, GS_ID(DETAIL_LIGHTING_CONTRAST_MEDIUM), },
-    { LIGHTING_CONTRAST_HIGH, GS_ID(DETAIL_LIGHTING_CONTRAST_HIGH), },
+    { LIGHTING_CONTRAST_LOW, GS_ID(GRAPHIC_SETTINGS_LIGHTING_CONTRAST_LOW), },
+    { LIGHTING_CONTRAST_MEDIUM, GS_ID(GRAPHIC_SETTINGS_LIGHTING_CONTRAST_MEDIUM), },
+    { LIGHTING_CONTRAST_HIGH, GS_ID(GRAPHIC_SETTINGS_LIGHTING_CONTRAST_HIGH), },
     { -1, nullptr, },
     // clang-format on
 };
 
 static const UI_SETTINGS_ENUM_ENTRY m_RenderModeEnumEntries[] = {
     // clang-format off
-    { RM_SOFTWARE, GS_ID(DETAIL_RENDER_MODE_SOFTWARE), },
-    { RM_HARDWARE, GS_ID(DETAIL_RENDER_MODE_HARDWARE), },
+    { RM_SOFTWARE, GS_ID(GRAPHIC_SETTINGS_RENDER_MODE_SOFTWARE), },
+    { RM_HARDWARE, GS_ID(GRAPHIC_SETTINGS_RENDER_MODE_HARDWARE), },
     { -1, nullptr, },
     // clang-format on
 };
 
 static const UI_SETTINGS_ENUM_ENTRY m_AspectModeEnumEntries[] = {
     // clang-format off
-    { AM_4_3, GS_ID(DETAIL_ASPECT_MODE_4_3), },
-    { AM_16_9, GS_ID(DETAIL_ASPECT_MODE_16_9), },
-    { AM_ANY, GS_ID(DETAIL_ASPECT_MODE_ANY), },
+    { AM_4_3, GS_ID(GRAPHIC_SETTINGS_ASPECT_MODE_4_3), },
+    { AM_16_9, GS_ID(GRAPHIC_SETTINGS_ASPECT_MODE_16_9), },
+    { AM_ANY, GS_ID(GRAPHIC_SETTINGS_ASPECT_MODE_ANY), },
     { -1, nullptr, },
     // clang-format on
 };
@@ -45,7 +45,7 @@ static const UI_SETTINGS_ENUM_ENTRY m_AspectModeEnumEntries[] = {
 static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
     {
         .option_type = COT_INT32,
-        .label_id = GS_ID(DETAIL_FOG_START),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_FOG_START),
         .target = &g_Config.visuals.fog_start,
         .min_value = 1,
         .max_value = 100,
@@ -53,7 +53,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
 
     {
         .option_type = COT_INT32,
-        .label_id = GS_ID(DETAIL_FOG_END),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_FOG_END),
         .target = &g_Config.visuals.fog_end,
         .min_value = 1,
         .max_value = 100,
@@ -61,7 +61,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
 
     {
         .option_type = COT_RGB888,
-        .label_id = GS_ID(DETAIL_WATER_COLOR_R),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_WATER_COLOR_R),
         .target = &g_Config.visuals.water_color,
         .min_value = 0,
         .max_value = 255,
@@ -72,7 +72,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
 
     {
         .option_type = COT_RGB888,
-        .label_id = GS_ID(DETAIL_WATER_COLOR_G),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_WATER_COLOR_G),
         .target = &g_Config.visuals.water_color,
         .min_value = 0,
         .max_value = 255,
@@ -83,7 +83,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
 
     {
         .option_type = COT_RGB888,
-        .label_id = GS_ID(DETAIL_WATER_COLOR_B),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_WATER_COLOR_B),
         .target = &g_Config.visuals.water_color,
         .min_value = 0,
         .max_value = 255,
@@ -94,7 +94,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
 
     {
         .option_type = COT_ENUM,
-        .label_id = GS_ID(DETAIL_CAMERA_MODE),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_CAMERA_MODE),
         .target = &g_Config.visuals.camera_mode,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -103,7 +103,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
 
     {
         .option_type = COT_INT32,
-        .label_id = GS_ID(DETAIL_FOV),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_FOV),
         .target = &g_Config.visuals.fov,
         .min_value = 30,
         .max_value = 150,
@@ -113,7 +113,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
 
     {
         .option_type = COT_BOOL,
-        .label_id = GS_ID(DETAIL_USE_PSX_FOV),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_USE_PSX_FOV),
         .target = &g_Config.visuals.use_psx_fov,
     },
 
@@ -125,7 +125,7 @@ static const UI_SETTINGS_OPTION m_VisualsOptions[] = {
 static const UI_SETTINGS_OPTION m_UIOptions[] = {
     {
         .option_type = COT_DOUBLE,
-        .label_id = GS_ID(DETAIL_UI_TEXT_SCALE),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_UI_TEXT_SCALE),
         .target = &g_Config.ui.text_scale,
         .min_value = 50,
         .max_value = 200,
@@ -135,7 +135,7 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
 
     {
         .option_type = COT_DOUBLE,
-        .label_id = GS_ID(DETAIL_UI_BAR_SCALE),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_UI_BAR_SCALE),
         .target = &g_Config.ui.bar_scale,
         .min_value = 50,
         .max_value = 200,
@@ -145,7 +145,7 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
 
     {
         .option_type = COT_BOOL,
-        .label_id = GS_ID(DETAIL_UI_SCROLL_WRAPAROUND),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_UI_SCROLL_WRAPAROUND),
         .target = &g_Config.ui.enable_wraparound,
     },
 
@@ -157,7 +157,7 @@ static const UI_SETTINGS_OPTION m_UIOptions[] = {
 static const UI_SETTINGS_OPTION m_RenderOptions[] = {
     {
         .option_type = COT_INT32,
-        .label_id = GS_ID(DETAIL_FPS),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_FPS),
         .target = &g_Config.rendering.fps,
         .min_value = 30,
         .max_value = 60,
@@ -167,7 +167,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
 
     {
         .option_type = COT_ENUM,
-        .label_id = GS_ID(DETAIL_TEXTURE_FILTER),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_TEXTURE_FILTER),
         .target = &g_Config.rendering.texture_filter,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -176,19 +176,19 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
 
     {
         .option_type = COT_BOOL,
-        .label_id = GS_ID(DETAIL_TRAPEZOID_FILTER),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_TRAPEZOID_FILTER),
         .target = &g_Config.rendering.enable_trapezoid_filter,
     },
 
     {
         .option_type = COT_BOOL,
-        .label_id = GS_ID(DETAIL_DEPTH_BUFFER),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_DEPTH_BUFFER),
         .target = &g_Config.rendering.enable_zbuffer,
     },
 
     {
         .option_type = COT_ENUM,
-        .label_id = GS_ID(DETAIL_LIGHTING_CONTRAST),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_LIGHTING_CONTRAST),
         .target = &g_Config.rendering.lighting_contrast,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -197,7 +197,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
 
     {
         .option_type = COT_ENUM,
-        .label_id = GS_ID(DETAIL_RENDER_MODE),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_RENDER_MODE),
         .target = &g_Config.rendering.render_mode,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -206,7 +206,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
 
     {
         .option_type = COT_ENUM,
-        .label_id = GS_ID(DETAIL_ASPECT_MODE),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_ASPECT_MODE),
         .target = &g_Config.rendering.aspect_mode,
         .delta_slow = 1,
         .delta_fast = 1,
@@ -215,7 +215,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
 
     {
         .option_type = COT_INT32,
-        .label_id = GS_ID(DETAIL_SCALER),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_SCALER),
         .target = &g_Config.rendering.scaler,
         .min_value = 1,
         .max_value = 4,
@@ -225,7 +225,7 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
 
     {
         .option_type = COT_FLOAT,
-        .label_id = GS_ID(DETAIL_SIZER),
+        .label_id = GS_ID(GRAPHIC_SETTINGS_SIZER),
         .target = &g_Config.rendering.sizer,
         .min_value = 40,
         .max_value = 200,
@@ -239,15 +239,16 @@ static const UI_SETTINGS_OPTION m_RenderOptions[] = {
 };
 
 static const UI_SETTINGS_TAB m_Tabs[] = {
-    { GS_ID(DETAIL_VISUALS_TAB), m_VisualsOptions },
-    { GS_ID(DETAIL_UI_TAB), m_UIOptions },
-    { GS_ID(DETAIL_RENDERING_TAB), m_RenderOptions },
+    { GS_ID(GRAPHIC_SETTINGS_VISUALS_TAB), m_VisualsOptions },
+    { GS_ID(GRAPHIC_SETTINGS_UI_TAB), m_UIOptions },
+    { GS_ID(GRAPHIC_SETTINGS_RENDERING_TAB), m_RenderOptions },
 };
 
 void UI_GraphicSettings_Init(UI_GRAPHIC_SETTINGS_STATE *const s)
 {
     UI_Settings_InitWithTabs(
-        s, GS_ID(DETAIL_TITLE), sizeof(m_Tabs) / sizeof(m_Tabs[0]), m_Tabs);
+        s, GS_ID(GRAPHIC_SETTINGS_TITLE), sizeof(m_Tabs) / sizeof(m_Tabs[0]),
+        m_Tabs);
 }
 
 void UI_GraphicSettings_Free(UI_GRAPHIC_SETTINGS_STATE *const s)